Implement SSR for ResponsiveBox
