+++ testing/tests/DevExpress.serverSide/responsiveBox.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+import "../DevExpress.ui.widgets/responsiveBox.markup.tests.js";
